Extract StatusCommand to its own file
